+++ src/Squidlr.App/Pages/MainPage.xaml
@@ -1,9 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:local="clr-namespace:Squidlr.App"
-             x:Class="Squidlr.App.MainPage"
-             x:DataType="local:MainPageViewModel">
+             xmlns:pages="clr-namespace:Squidlr.App.Pages"
+             x:Class="Squidlr.App.Pages.MainPage"
+             x:DataType="pages:MainPageViewModel">
 
     <ScrollView>
         <VerticalStackLayout
